pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "japon"
version = "0.1.0"
description = "Planificador de viaje a Japón 2027 — monitor de vuelos BUE→TYO"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13.7",
    "fast-flights>=0.2",
    "amadeus>=9.0",
    "apscheduler>=3.10",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.3",
    "httpx>=0.27",
]

[project.scripts]
japon = "japon.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/japon"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
]
